requirements.txt
discord.py
pytz
emoji